# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Internet Bank Backend API
  version: 1.1.0
  description: |
    Актуальная HTTP-спецификация backend-сервисов интернет-банка через API Gateway.
    Все запросы отправляются на один базовый URL.
servers:
  - url: http://localhost:8080
    description: API Gateway
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, type, roles, email, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [client, employee]
        roles:
          type: array
          items:
            type: string
            enum: [client, employee]
        email:
          type: string
          format: email
        full_name:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, blocked]
        created_at:
          type: string
          format: date-time
...
    UserListResponse:
      type: object
      required: [users, page_number, page_quantity]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        page_number:
          type: integer
        page_quantity:
          type: integer
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [client, employee]
          description: Необязательная роль для выпуска токена при multi-role аккаунте.
    LoginResponse:
      type: object
      required: [token, refresh_token, user_id, type]
      properties:
        token:
          type: string
        refresh_token:
          type: string
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [client, employee]
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    CreateUserRequest:
      type: object
      required: [type, email, full_name]
      properties:
        type:
          type: string
          enum: [client, employee]
        email:
          type: string
          format: email
        full_name:
          type: string
        phone:
          type: string
        password:
          type: string
          description: Если не указан, используется пароль по умолчанию.
...
    SSOTokenResponse:
      type: object
      required: [token_type, access_token, token, refresh_token, expires_in, user_id, type]
      properties:
        token_type:
          type: string
          example: Bearer
        access_token:
          type: string
        token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          example: 900
        user_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [client, employee]
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Авторизация пользователя
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Неверное тело запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверный логин/пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление пары токенов
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Неверное тело запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Невалидный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sso/authorize:
    get:
      tags: [SSO]
      summary: Запуск OAuth2 Authorization Code + PKCE
      security: []
      parameters:
        - in: query
          name: response_type
          required: true
          schema: { type: string, example: code }
        - in: query
          name: client_id
          required: true
          schema: { type: string }
        - in: query
          name: redirect_uri
          required: true
          schema: { type: string, format: uri }
        - in: query
          name: state
          required: false
          schema: { type: string }
        - in: query
          name: code_challenge
          required: true
          schema: { type: string }
        - in: query
          name: code_challenge_method
          required: true
          schema: { type: string, example: S256 }
      responses:
        '302':
          description: Redirect на страницу логина или в redirect_uri с code
  /sso/login:
    get:
      tags: [SSO]
      summary: HTML-форма входа для SSO
      security: []
      parameters:
        - in: query
          name: response_type
          required: true
          schema: { type: string }
        - in: query
          name: client_id
          required: true
          schema: { type: string }
        - in: query
          name: redirect_uri
          required: true
          schema: { type: string, format: uri }
        - in: query
          name: state
          required: false
          schema: { type: string }
        - in: query
          name: code_challenge
          required: true
          schema: { type: string }
        - in: query
          name: code_challenge_method
          required: true
          schema: { type: string }
      responses:
        '200':
          description: HTML-страница
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [SSO]
      summary: Авторизация пользователя на SSO-форме
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email, password, response_type, client_id, redirect_uri, code_challenge, code_challenge_method]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                response_type: { type: string }
                client_id: { type: string }
                redirect_uri: { type: string, format: uri }
                state: { type: string }
                code_challenge: { type: string }
                code_challenge_method: { type: string }
      responses:
        '302':
          description: Redirect в redirect_uri с authorization code
  /sso/token:
    post:
      tags: [SSO]
      summary: Обмен authorization code на access/refresh токены
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type, code, client_id, redirect_uri, code_verifier]
              properties:
                grant_type:
                  type: string
                  example: authorization_code
                code:
                  type: string
                client_id:
                  type: string
                redirect_uri:
                  type: string
                code_verifier:
                  type: string
      responses:
        '200':
          description: Токены выданы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOTokenResponse'
        '400':
          description: Неверные параметры
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Код недействителен / верификация PKCE не прошла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    post:
      tags: [Users]
      summary: Создать пользователя (только сотрудник)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
        '403':
          description: Только для сотрудника
    get:
      tags: [Users]
      summary: Список пользователей
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [client, employee]
        - in: query
          name: status
          schema:
            type: string
            enum: [active, blocked]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 100
            minimum: 1
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Получить пользователя по ID
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Users]
      summary: Блокировка/разблокировка пользователя
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: action
          required: true
          schema:
            type: string
            enum: [block, unblock]
      responses:
        '204':
          description: Статус обновлен
        '403':
          description: Нельзя заблокировать себя / нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...